+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 ### New features
 
 * New config setting `:libspec-whitelist` which makes it possible to create a seq of namespaces `clean-ns` shouldn't prune.  This is useful for libspecs which aren't used except through side-effecting loads.
+* New config setting `:ignore-paths` for ignoring certain paths when finding dirs on classpath.
 
 ## 2.2.0